Add Assassin's Creed: Brotherhood - Sequence 05 Completed
Throughout Sequence 05 of Assassin's Creed: Brotherhood, Ezio’s efforts to dismantle the financial stronghold of the Borgia family by targeting The Banker. Ezio successfully infiltrates a party to assassinate The Banker and disrupts the Borgia’s funding, which is crucial for their control over Rome. This sequence is pivotal as it weakens the Borgia’s influence and showcases Ezio’s strategic approach to liberating the city. It’s time to join the Brotherhood. Here's a brief overview:

1. Escape from Debt: Ezio must escape from debt collectors after a failed investment.

2. Follow The Money: Ezio tracks the flow of money to uncover the identity of The Banker.

3. When In Rome…: Disguised as a guard, Ezio must deliver a chest of money to The Banker.

4. In And Out: Ezio attends a party to assassinate The Banker.

5. Paper Trail: After eliminating The Banker, Ezio must escape and avoid being caught by the guards.

In addition to these main memories, there are special missions as well

The Da Vinci Disappearance - Restored Memory:

1. A Roll of The Dice: Ezio must find Leonardo’s assistant, Salai, and bring him back to Leonardo’s workshop while avoiding fights and ensuring Salai’s safety4.

Cristina Mission:

1. Last Rites: Ezio has repressed memories of his first love, Cristina Vespucci. In this memory, he must lead Cristina to safety and transport bodies to a boat without being detected5.

War Machine:

1. War Plans: Ezio kills the Templar Overseer to recover a map showing the location of one of Leonardo’s war machines.

2. Loose Cannon: The Borgia have built a deadly Naval War Machine from Leonardo’s designs. Ezio must destroy it to prevent Cesare Borgia from controlling the seas.